requirements.txt
pandas
openpyxl
requests
python-dotenv
langchain-core
langchain-openai
langgraph
typing-extensions
langchain-community
langchain-experimental
networkx
langchain
pydantic
pydantic-settings
pydantic_core
langchain-text-splitters
jsonschema
azure-identity
azure-ai-documentintelligence
azure-search-documents
tiktoken 
requests
azure-identity
azure-core
azure-search-documents
azure-ai-documentintelligence
pyvis
azure-ai-translation-text
azure-storage-blob
python-docx
python-pptx
docx2pdf
azure-storage-file-datalake
PyPDF2
pyodbc
databricks-sql-connector